chore: rename to 'books-route'
