+++ src/app/(main)/books/[isbn]/loading.tsx
@@ -1,4 +1,4 @@
-import { BookPageLoadingSkeleton } from "@/client/components/modules/book-page/book-page-loading-skeleton";
+import { BookPageLoadingSkeleton } from "@/client/components/modules/books-route/book-page-loading-skeleton";
 
 export default function BookPageLoading() {
   return <BookPageLoadingSkeleton />;
