+++ src/app/(main)/books/[isbn]/page.tsx
@@ -1,4 +1,4 @@
-import { BookPageWrapper } from "@/root/src/libs/client/src/components/modules/book-page/book-page-wrapper";
+import { BookPageWrapper } from "@/root/src/libs/client/src/components/modules/books-route/book-page-wrapper";
 
 export default function Book({ params: { isbn } }: { params: { isbn: string } }) {
   return <BookPageWrapper params={{ isbn }} />;
