+++ src/libs/client/src/components/modules/books-route/book-information.tsx
