+++ src/libs/client/src/components/modules/books-route/book-page-loading-skeleton.tsx
